+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [6.5] - 2025-09-30
+## [6.5.0] - 2025-09-30
 
 ### Évolutions
 - **Intégration Sentry** (Redmine 26581)